agent_os/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
playwright>=1.44.0
python-dotenv>=1.0.0
rich>=13.7.0
pydantic>=2.7.0
